Support using .cfformat.json file by convention
Starting with the directory formatting will take place in, search
there and in every parent directory for a .cfformat.json file until
one is found or the root directory is reached. If one is found it
will be used. If a `.git` folder is encountered, this will also
stop the search.